Combine cart and checkout stores
